Revising tests of MNGM
